refactor(auth-context): move AuthContext from js to jsx
