+++ src/index.jsx
@@ -2,7 +2,7 @@ import React from "react";
 import ReactDOM from "react-dom";
 import "./index.scss";
 import App from "./App";
-import { AuthContextProvider } from "./context/auth/auth-context";
+import { AuthContextProvider } from "./context/auth/AuthContext";
 
 ReactDOM.render(
   <React.StrictMode>
